// source/frontend/src/models/api/NotificationSearchCriteria.ts
export interface Api_NotificationSearchCriteria {
  type?: string | null;
  acquisitionFileId?: number | null;
  dispositionFileId?: number | null;
  researchFileId?: number | null;
  managementFileId?: number | null;
  leaseId?: number | null;
  takeId?: number | null;
  insuranceId?: number | null;
  leaseConsultationId?: number | null;
  noticeOfClaimId?: number | null;
  leaseRenewalId?: number | null;
  expropOwnerHistoryId?: number | null;
  agreementId?: number | null;
}